requirements.txt
google-api-python-client==2.129.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
requests==2.32.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
google-generativeai>=0.7.2
